opply_project/requirements.txt
asgiref==3.7.2
Django==4.2.11
django-environ==0.11.2
django-filter==23.5
django-guid==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
execnet==2.0.2
importlib_metadata==7.0.2
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.5.2
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
zipp==3.17.0
